refactor: rename navigation to menu
