+++ src/_shell/menu/menu.config.tsx
@@ -3,7 +3,7 @@ import { ServiceIcon } from '../top-bar/icons/service-icon';
 import { TargetIcon } from '../top-bar/icons/target-icon';
 import type { MenuConfig } from './types';
 
-export const navigationConfig: MenuConfig = {
+export const menuConfig: MenuConfig = {
   items: [
     {
       text: 'Für Unternehmen',
