+++ src/_shell/menu/mobile/accordion/collapsible-accordion-item.tsx
