+++ src/_shell/menu/mobile/icons/cross.icon.tsx
