+++ src/_shell/menu/mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './mobile-menu';
