+++ src/_shell/menu/types/entry-menu-item.type.ts
