--- src/_shell/navigation/mobile/index.ts
@@ -1,1 +1,0 @@
-export * from './mobile-navigation';
